Extract sort and filter configs to constants
